mmssms_merge: initial script structure
Add the initial mmssms_merge.sh script with its intended basic
structure.
It is supposed to run on a rooted Android phone to merge a different
mmssms.db SQlite database with the one in use on the phone.
It is certainly a good idea to do a (Nandroid-)backup of the phone data
first, and disable wireless networks (airplane mode) from the time of
the backup until the script has been successfully finished, so that
incoming messages will not interfere.